// src/features/cross-chain/providers/bridgers-provider/tron-bridgers-trade/models/tron-bridgers-transaction-data.ts
import { TronParameters } from 'src/core/blockchain/web3-pure/typed-web3-pure/tron-web3-pure/models/tron-parameters';

export interface TronBridgersTransactionData {
    functionName: string;
    options: {
        feeLimit: number;
        callValue: string;
    };
    parameter: TronParameters;
    to: string;
}